+++ src/components/app/App.js
@@ -1,22 +1,9 @@
-import { useState } from "react";
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-
+import {MainPage, ComicsPage} from '../pages';
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import ComicsList from "../comicsList/ComicsList";
-
-import decoration from '../../resources/img/vision.png';
-import AppBanner from "../appBanner/AppBanner";
 
 const App = () => {
-    const [selectedChar, setChar] = useState(null);
 
-    const onCharSelected = (id) => {
-        setChar(id);
-    }
 
     return (
         <Router>
@@ -25,22 +12,10 @@ const App = () => {
                 <main>
                     <Switch>
                         <Route exact path="/">
-                            <ErrorBoundary>
-                                <RandomChar/>
-                            </ErrorBoundary>
-                            <div className="char__content">
-                                <ErrorBoundary>
-                                    <CharList onCharSelected={onCharSelected}/>
-                                </ErrorBoundary>
-                                <ErrorBoundary>
-                                    <CharInfo charId={selectedChar}/>
-                                </ErrorBoundary>
-                            </div>
-                            <img className="bg-decoration" src={decoration} alt="vision"/>
+                            <MainPage/>
                         </Route>
                         <Route exact path="/comics">
-                            <AppBanner/>
-                            <ComicsList/>
+                            <ComicsPage/>
                         </Route>
                     </Switch>
                 </main>
